End of Section 5

+++ AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -41,13 +42,14 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
         return mapper.Map<AuctionDto>(auction);
     }
-
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = mapper.Map<Auction>(createAuctionDto);
-        // TODO: Add the current user to the auction as Seller
-        auction.Seller = "test";
+        
+
+        auction.Seller = User.Identity?.Name ?? "Unknown user";
 
         context.Auctions.Add(auction);
 
@@ -63,6 +65,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
             new { auction.Id }, mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -71,7 +74,8 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -87,7 +91,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
         return BadRequest("Problem saving changes");
     }
-
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -95,7 +99,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper, IPubli
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         context.Auctions.Remove(auction);
